python-requirements.txt
# Copyright lowRISC contributors.
# Licensed under the Apache License, Version 2.0, see LICENSE for details.
# SPDX-License-Identifier: Apache-2.0

# Keep sorted.
bokeh==2.4.3
cffi==1.16.0
codetiming==1.4.0
gitpython==3.1.40
libusb1==3.1.0
matplotlib==3.8.1
more-itertools==10.1.0
networkx==3.2.1
pycryptodome==3.19.0
# TODO: Use pyvisa instead after switching to a dedicated PRNG
python-vxi11==0.9
pyyaml==6.0.1
ray==2.8.0
scared==1.1.0
scipy==1.11.3
sqlalchemy==2.0.23
sqlalchemy_utils==0.41.1
tqdm==4.66.1
trsfile==2.2.0
typer==0.9.0
wheel==0.41.3
# can be removed after switching to ray
joblib==1.3.2
gitpython==3.1.40

# Development version of ChipWhisperer toolchain with latest features and
# bug fixes - Needs to be installed in editable mode. We fix the version
# for improved stability and manually update if necessary.
-e git+https://github.com/newaetech/chipwhisperer-historical.git@3eace1719daf43d4f0965c1790c2c8a9e8b2f690#egg=chipwhisperer

# Development version of ChipShover toolchain with latest features and
# bug fixes - Needs to be installed in editable mode. We fix the version
# for improved stability and manually update if necessary.
-e git+https://github.com/newaetech/ChipShover.git@f824426668b2d1f74367fdb7071058d27b17ded4#egg=chipshover&subdirectory=host-python

# # Development version of ChipShouter toolchain with latest features and
# # bug fixes - Needs to be installed in editable mode. We fix the version
# # for improved stability and manually update if necessary.
-e git+https://github.com/newaetech/ChipSHOUTER-python.git@9ba55f5b12c746fcb31d675b4784d5f339760965#egg=chipshouter

# Linters
-r python-requirements-lint.txt

# Test-only requirements
pytest==7.4.3
